+++ ci
@@ -9,6 +9,9 @@ for ex in examples/*.py; do
     PYTHONPATH=. uv run "$ex"
 done
 
+# Unit testing runs the queue tests
+uv run -m unittest -v jobserver.test_queue
+
 # Unit testing runs the executor tests
 uv run -m unittest -v jobserver.test_executor
