+++ jobserver/_jobserver.py
@@ -15,11 +15,11 @@ import types
 from typing import Any, Generic, NoReturn, Optional, TypeVar, Union
 
 # Implementation depends upon an explicit subset of multiprocessing
-from multiprocessing import get_context
 from multiprocessing.connection import Connection, wait
 from multiprocessing.context import BaseContext
 from multiprocessing.process import BaseProcess
-from multiprocessing.reduction import ForkingPickler  # type: ignore
+
+from ._queue import MinimalQueue, absolute_deadline, resolve_context
 
 __all__ = (
     "Blocked",
@@ -247,126 +247,6 @@ class Future(Generic[T]):
 
         assert self._wrapper is not None
         return self._wrapper.unwrap()
-
-
-def absolute_deadline(relative_timeout: Optional[float]) -> float:
-    """
-    Convert relative_timeout in seconds into a monotonic, absolute deadline.
-
-    A large, absolute timeout is returned whenever relative_timeout is None.
-    """
-    # Cannot be Inf nor sys.float_info.max nor sys.maxsize / 1000
-    # nor _PyTime_t per https://stackoverflow.com/questions/45704243!
-    return time.monotonic() + (
-        (60 * 60 * 24 * 7)  # 604.8k seconds ought to be enough for anyone
-        if relative_timeout is None
-        else relative_timeout
-    )
-
-
-def resolve_context(context: Union[None, str, BaseContext]) -> BaseContext:
-    """Return a multiprocessing BaseContext, resolving None/str as needed."""
-    if context is None or isinstance(context, str):
-        return get_context(context)
-    return context
-
-
-class MinimalQueue(Generic[T]):
-    """
-    An unbounded SimpleQueue-variant with minimal function needed by Jobserver.
-
-    Vanilla multiprocessing.SimpleQueue lacks timeout on get(...).
-    Vanilla multiprocessing.Queue has wildly undesired threading machinery.
-    Both get(...) and put(...) detect and report when one end hangs up.
-    """
-
-    __slots__ = ("_reader", "_writer", "_read_lock", "_write_lock")
-
-    def __init__(self, context: Union[None, str, BaseContext] = None) -> None:
-        """Use given context with default of multiprocessing.get_context()."""
-        context = resolve_context(context)
-        reader, writer = context.Pipe(duplex=False)
-        self._reader: Optional[Connection] = reader
-        self._writer: Optional[Connection] = writer
-        self._read_lock = context.Lock()
-        self._write_lock = context.Lock()
-
-    def __copy__(self) -> "MinimalQueue":
-        """Shallow copies return the original MinimalQueue unchanged."""
-        # Because any "copy" should and can only mutate same pipe/locks
-        return self
-
-    def __deepcopy__(self, _: Any) -> "MinimalQueue":
-        """Deep copies return the original MinimalQueue unchanged."""
-        # Because any "copy" should and can only mutate same pipe/locks
-        return self
-
-    def waitable(self) -> int:
-        """The object on which to wait(...) to get(...) new data."""
-        assert self._reader is not None, "waitable() after close_get()"
-        return self._reader.fileno()
-
-    def close_get(self) -> None:
-        """Close the receiving end; get() may no longer be called.
-
-        Closes the underlying pipe end so EOF propagates on crash.
-        """
-        if self._reader is not None:
-            self._reader.close()
-            self._reader = None
-            # _read_lock must NOT be set to None here.  In spawn/forkserver
-            # mode, setting it to None would drop the last parent-side
-            # reference, causing CPython to call sem_unlink() on the named
-            # semaphore before the child process has had a chance to open it,
-            # producing FileNotFoundError during the child's unpickling.
-
-    def close_put(self) -> None:
-        """Close the sending end; put() may no longer be called.
-
-        Closes the underlying pipe end so EOF propagates on crash.
-        """
-        if self._writer is not None:
-            self._writer.close()
-            self._writer = None
-            # _write_lock must NOT be set to None here; same race condition
-            # as described in close_get().
-
-    def get(self, timeout: Optional[float] = None) -> T:
-        """
-        Get one object from the queue raising queue.Empty if unavailable.
-
-        Raises EOFError on exhausted queue whenever sending half has hung up.
-        """
-        assert self._reader is not None, "get() after close_get()"
-        # Accounting for lock acquisition time is easiest with a deadline
-        # and conditionals repeatedly checking for negative situations
-        # Otherwise, this turns into an unpleasantly messy stretch of code
-        deadline = absolute_deadline(relative_timeout=timeout)
-        if not self._read_lock.acquire(block=True, timeout=timeout):
-            raise queue.Empty
-        try:
-            if not self._reader.poll(deadline - time.monotonic()):
-                raise queue.Empty
-            recv = self._reader.recv_bytes()
-        finally:
-            self._read_lock.release()
-
-        # Deserialize outside the critical section
-        return ForkingPickler.loads(recv)
-
-    def put(self, *args: T) -> None:
-        """
-        Put zero or more objects into the queue, contiguously.
-
-        Raises BrokenPipeError if the receiving half has hung up.
-        """
-        assert self._writer is not None, "put() after close_put()"
-        if args:
-            # Serialize outside the critical section
-            send = [ForkingPickler.dumps(arg) for arg in args]
-            with self._write_lock:
-                while send:
-                    self._writer.send_bytes(send.pop(0))
 
 
 # Appears as a default argument in Jobserver to simplify some logic therein
