+++ jobserver/test_executor.py
@@ -25,7 +25,8 @@ from multiprocessing import get_all_start_methods
 
 from ._executor import JobserverExecutor
 from ._request import Submit
-from ._jobserver import Jobserver, MinimalQueue
+from ._jobserver import Jobserver
+from ._queue import MinimalQueue
 
 # Most tests use the fastest start method.  On Python 3.12+ "fork" is
 # deprecated when the process is multi-threaded, so fall back to
