+++ jobserver/test_jobserver.py
@@ -23,9 +23,9 @@ from ._jobserver import (
     CallbackRaised,
     Future,
     Jobserver,
-    MinimalQueue,
     SubmissionDied,
 )
+from ._queue import MinimalQueue
 
 T = typing.TypeVar("T")
 
@@ -207,34 +207,6 @@ class JobserverTest(unittest.TestCase):
                 # copy.copy(...) and copy.deepcopy(...) return the original.
                 self.assertIs(js1, js2)
                 self.assertIs(js1, js3)
-
-    # No behavioral assertions made around pickling, however.
-    # Indeed, disabling pickling MinimalQueues empirically breaks many tests.
-    def test_duplication_minimalqueue(self) -> None:
-        """Copying of MinimalQueue is explicitly allowed."""
-        for method in get_all_start_methods():
-            with self.subTest(method=method):
-                mq1 = MinimalQueue(context=method)  # type: MinimalQueue[int]
-                mq2 = copy.copy(mq1)
-                mq3 = copy.deepcopy(mq1)
-                mq1.put(1)
-                mq2.put(2)
-                mq3.put(3)
-                self.assertEqual(1, mq3.get())
-                self.assertEqual(2, mq2.get())
-                self.assertEqual(3, mq1.get())
-                # Though copying is allowed, it is degenerate in that
-                # copy.copy(...) and copy.deepcopy(...) return the original.
-                self.assertIs(mq1, mq2)
-                self.assertIs(mq1, mq3)
-
-    def test_close_get_and_close_put_are_idempotent(self) -> None:
-        """close_get() and close_put() are safe to call more than once."""
-        mq: MinimalQueue[int] = MinimalQueue()
-        mq.close_get()
-        mq.close_get()  # second call must not raise
-        mq.close_put()
-        mq.close_put()  # second call must not raise
 
     # Motivated by multiprocessing.Connection mentioning a possible 32MB limit
     def test_large_objects(self) -> None:
